+++ core-ng/src/main/java/core/framework/impl/cache/RedisCacheStore.java
@@ -1,11 +1,12 @@
 package core.framework.impl.cache;
 
+import core.framework.impl.redis.RedisException;
 import core.framework.impl.redis.RedisImpl;
 import core.framework.util.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import redis.clients.jedis.exceptions.JedisConnectionException;
 
+import java.io.UncheckedIOException;
 import java.time.Duration;
 import java.util.Map;
 
@@ -25,7 +26,7 @@ public class RedisCacheStore implements CacheStore {
     public byte[] get(String key) {
         try {
             return redis.getBytes(key);
-        } catch (JedisConnectionException e) {
+        } catch (UncheckedIOException | RedisException e) {
             logger.warn("failed to connect to redis, error={}", e.getMessage(), e);
             return null;
         }
@@ -35,7 +36,7 @@ public class RedisCacheStore implements CacheStore {
     public Map<String, byte[]> getAll(String[] keys) {
         try {
             return redis.multiGetBytes(keys);
-        } catch (JedisConnectionException e) {
+        } catch (UncheckedIOException | RedisException e) {
             logger.warn("failed to connect to redis, error={}", e.getMessage(), e);
             return Maps.newHashMap();
         }
@@ -45,7 +46,7 @@ public class RedisCacheStore implements CacheStore {
     public void put(String key, byte[] value, Duration expiration) {
         try {
             redis.set(key, value, expiration);
-        } catch (JedisConnectionException e) {
+        } catch (UncheckedIOException | RedisException e) {
             logger.warn("failed to connect to redis, error={}", e.getMessage(), e);
         }
     }
@@ -54,7 +55,7 @@ public class RedisCacheStore implements CacheStore {
     public void putAll(Map<String, byte[]> values, Duration expiration) {
         try {
             redis.multiSet(values, expiration);
-        } catch (JedisConnectionException e) {
+        } catch (UncheckedIOException | RedisException e) {
             logger.warn("failed to connect to redis, error={}", e.getMessage(), e);
         }
     }
@@ -63,7 +64,7 @@ public class RedisCacheStore implements CacheStore {
     public void delete(String key) {
         try {
             redis.del(key);
-        } catch (JedisConnectionException e) {
+        } catch (UncheckedIOException | RedisException e) {
             logger.warn("failed to connect to redis, error={}", e.getMessage(), e);
         }
     }