+++ core-ng/src/main/java/core/framework/impl/redis/RedisException.java
@@ -1,4 +1,4 @@
-package core.framework.impl.redis.v2;
+package core.framework.impl.redis;
 
 /**
  * @author neo